- Replace @Service and @Bean with functional approach - Add profiles: istio, keycloak - Bump dependencies
