+++ src/main/kotlin/org/izmaylovalexey/services/istio/IstioRole.kt
@@ -17,9 +17,7 @@ import org.izmaylovalexey.services.RoleService
 import org.izmaylovalexey.services.RoleTemplateService
 import org.izmaylovalexey.services.Success
 import org.izmaylovalexey.services.toFailure
-import org.springframework.stereotype.Service
 
-@Service
 internal class IstioRole(val istioClient: IstioClient) : RoleService, RoleTemplateService {
 
     private val mapper = YAMLMapper().registerKotlinModule()
