+++ src/main/kotlin/org/izmaylovalexey/services/keycloak/KeycloakTenant.kt
@@ -31,10 +31,8 @@ import org.izmaylovalexey.services.TenantService
 import org.izmaylovalexey.services.toFailure
 import org.keycloak.admin.client.Keycloak
 import org.keycloak.representations.idm.GroupRepresentation
-import org.springframework.stereotype.Service
 import javax.ws.rs.NotFoundException
 
-@Service
 internal class KeycloakTenant(
     private val keycloak: Keycloak,
     keycloakProperties: KeycloakProperties,
