+++ src/main/resources/application-keycloak.yml
@@ -1,0 +1,12 @@
+keycloak:
+  uri: ${KEYCLOAK_URI:http://keycloak-http:80/auth}
+  realm: ${KEYCLOAK_REALM:istio}
+  client: ${KEYCLOAK_CLIENT:frontend}
+  secret: ${KEYCLOAK_SECRET:keycloak-http}
+  secretKey: ${KEYCLOAK_SECRET_KEY:password
+
+initial:
+  user:
+    email: ${INIT_USER_EMAIL}
+    password: ${INIT_USER_PASSWORD}
+    role: ${INIT_USER_ROLE}
