+++ src/test/resources/application-test.yml
@@ -1,0 +1,4 @@
+management:
+  endpoint:
+    shutdown:
+      enabled: true
