--- src/test/resources/application.yml
@@ -1,35 +1,0 @@
-server:
-  port: 8080
-  contextPath: /
-  http2:
-    enabled: true
-
-management:
-  endpoint:
-    metrics:
-      enabled: true
-    prometheus:
-      enabled: true
-    shutdown:
-      enabled: true
-  endpoints:
-    web:
-      exposure:
-        include: '*'
-  metrics:
-    export:
-      prometheus:
-        enabled: true
-
-tenant:
-  defaultRole: ${DEFAULT_ROLE:owner}
-
-keycloak:
-  realm: ${KEYCLOAK_REALM:test}
-  client: ${KEYCLOAK_CLIENT:frontend}
-
-initial:
-  user:
-    email: ${INIT_USER_EMAIL:operator@email}
-    password: ${INIT_USER_PASSWORD:12345}
-    role: ${INIT_USER_ROLE:operator}
